📃 docs: add dual MIT/Apache-2.0 license
This commit adds a standard dual-license scheme, allowing users to choose between the MIT and
Apache-2.0 licenses. It creates the `LICENSE-MIT` and `LICENSE-APACHE` files in the project root and
updates the `README.md` to reflect this change in the \"License\" and \"Contribution\" sections, as well
as in the license badge at the top of the file.